+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## `connpool` UNRELEASED
+
+- Added support for pulling items from the pool with waiting for the local limit.
+
 ## `connpool` 0.1.0 - December 28, 2025
 
 - First release